refactor(utils): move getTrackerName to utils
